adding util.py, refactor get language from views.py and parse g2p fallback from cli.py
