+++ readalongs/util.py
